reorganize package structure
