+++ simple_NER/annotators/datetime_ner.py
@@ -6,7 +6,7 @@ from lingua_franca.parse import extract_datetime, extract_duration
 from lingua_franca.lang.parse_en import _convert_words_to_numbers_en
 from lingua_franca.format import nice_duration, nice_date
 from lingua_franca import load_language
-from simple_NER.annotators.utils.diff import TextDiff
+from simple_NER.utils.diff import TextDiff
 
 
 load_language("en")
