+++ simple_NER/annotators/nltk_ner.py
@@ -3,7 +3,6 @@ from simple_NER import Entity
 
 try:
     import nltk
-
     nltk.download('punkt')
     nltk.download('averaged_perceptron_tagger')
     nltk.download('maxent_ne_chunker')
