+++ simple_NER/annotators/numbers_ner.py
@@ -2,7 +2,7 @@ from simple_NER.annotators import NERWrapper
 from simple_NER import Entity
 from lingua_franca.lang.parse_en import _convert_words_to_numbers_en
 from lingua_franca import load_language
-from simple_NER.annotators.utils.diff import TextDiff
+from simple_NER.utils.diff import TextDiff
 
 load_language("en")
