+++ simple_NER/res/en-us/FoxStoplist.txt
