+++ simple_NER/settings.py
@@ -1,3 +1,4 @@
 from os.path import join, dirname
 
-STOPLIST = join(dirname(__file__), "annotators", "utils", "keywords", "SmartStoplist.txt")
+STOPLIST = join(dirname(__file__), "res", "en-us", "SmartStoplist.txt")
+RESOURES_DIR = join(dirname(__file__), "res")
